Flytt klasser fra tilsyn barn til felles mapper Disse blir også bruk for beregning av læremidler
