+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/domain/BeregningsresultatTilsynBarn.kt
@@ -2,6 +2,7 @@ package no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.domain
 
 import no.nav.tilleggsstonader.kontrakter.felles.Periode
 import no.nav.tilleggsstonader.sak.felles.domain.BarnId
+import no.nav.tilleggsstonader.sak.vedtak.domain.Stønadsperiode
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.AktivitetType
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.MålgruppeType
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.Vilkårperiode
@@ -13,7 +14,6 @@ import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.YearMonth
 import java.util.UUID
-import no.nav.tilleggsstonader.sak.vilkår.stønadsperiode.domain.Stønadsperiode as StønadsperiodeInngangsvilkår
 
 data class BeregningsresultatTilsynBarn(
     val perioder: List<BeregningsresultatForMåned>,
@@ -54,18 +54,6 @@ data class StønadsperiodeGrunnlag(
     val antallDager: Int,
 )
 
-data class Stønadsperiode(
-    val id: UUID? = null,
-    override val fom: LocalDate,
-    override val tom: LocalDate,
-    val målgruppe: MålgruppeType,
-    val aktivitet: AktivitetType,
-) : Periode<LocalDate> {
-    init {
-        validatePeriode()
-    }
-}
-
 data class UtgiftBarn(
     val barnId: BarnId,
     val utgift: Int,
@@ -93,14 +81,3 @@ fun List<Vilkårperiode>.tilAktiviteter(): List<Aktivitet> {
             )
         }
 }
-
-fun StønadsperiodeInngangsvilkår.tilGrunnlagStønadsperiode() = Stønadsperiode(
-    id = this.id,
-    fom = this.fom,
-    tom = this.tom,
-    målgruppe = this.målgruppe,
-    aktivitet = this.aktivitet,
-)
-
-fun List<StønadsperiodeInngangsvilkår>.tilSortertGrunnlagStønadsperiode() =
-    this.map { it.tilGrunnlagStønadsperiode() }.sortedBy { it.fom }
